pyproject.toml
[project]
name = "economic-letter-app"
version = "0.1.0"
description = "Economic Letter Insights - AI-powered analysis of FRBSF economic letters"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "httpx>=0.25.0",
    "beautifulsoup4>=4.12.0",
    "boto3>=1.29.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "hypothesis>=6.92.0",
    "pytest-asyncio>=0.21.0",
]
